// src/shared/types/Tasks/TaskType.type.ts
export enum TaskBlockType {
  FILL_TEXT = 'FILL_TEXT',
  SEQUENCE = 'SEQUENCE',
  CHOOSE_OPTION = 'CHOOSE_OPTION',
  MATCH_PAIRS = 'MATCH_PAIRS',
  FREE_ANSWER = 'FREE_ANSWER',
  HIGHLIGHT_TEXT = 'HIGHLIGHT_TEXT',
  WORD_SCRAMBLE = 'WORD_SCRAMBLE',
  DIALOGUE = 'DIALOGUE',

  INFO_TEXT = 'INFO_TEXT',
  INFO_MEDIA = 'INFO_MEDIA',
  INFO_AUDIO = 'INFO_AUDIO'
}